backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
trafilatura==1.12.0
anthropic==0.34.2
python-dotenv==1.0.1
aiosqlite==0.20.0
